+++ include/RandLAPACK/comps/rf.hh
@@ -14,7 +14,7 @@ void rf1(
         int64_t p,
         int64_t passes_per_stab,
         T* Q, // n by k
-	uint64_t seed
+	uint32_t seed
 );
 
 } // end namespace RandLAPACK::comps::rs
